+++ lib/flatten.ts
@@ -1,8 +1,8 @@
 /**
- * **flatten** takes an array of elements and flattens it until all elements are on the root-level (`flat(Infinity)`)
+ * **flatten** takes an array of elements and flattens it by one level
  * @param xs array to be flattened
  * @returns a flattened array
  */
 export default function flatten<T>(xs: T[]): T[] {
-    return xs.flat(Infinity);
+    return xs.flat();
 }