+++ test/flatten.test.ts
@@ -7,7 +7,7 @@ test({
     name: "flatten",
     fn(): void {
         assertEquals(flatten([[1, 2, 3], [4, 5, 6]]), [1, 2, 3, 4, 5, 6]);
-        assertEquals(flatten([1, 2, [3], [4, [5, [6]]]]), [1, 2, 3, 4, 5, 6]);
+        assertEquals(flatten([1, 2, [3], [4, [5, [6]]]]), [1, 2, 3, 4, [5, [6]]]);
         assertEquals(flatten([]), []);
     }
 })