+++ CHANGELOG.md
@@ -6,7 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+
 ### Changed
+- flatten function flattens by one level instead of Infinity
+
 ### Removed
 
 ## [0.2.0] - 2020-02-06